app/requirements.txt
asgiref==3.6.0
Django==4.1.7
pip==22.3.1
setuptools==65.5.1
sqlparse==0.4.3
tzdata==2023.2
wheel==0.38.4
psycopg2==2.9.5
django-crispy-forms==2.0
crispy-bootstrap5==0.7
django-phonenumber-field==7.0.2
django-phonenumbers==1.0.1